requirements.txt
django==3.0.2
python-telegram-bot==12.2.0
celery[redis]
django-celery-results
django-celery-beat
